added cookies and csrf token

+++ AintBnB.WebApi/Controllers/AccommodationController.cs
@@ -1,14 +1,19 @@
-﻿using AintBnB.BusinessLogic.Interfaces;
+﻿using AintBnB.BusinessLogic.CustomExceptions;
+using AintBnB.BusinessLogic.Interfaces;
 using AintBnB.Core.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using static AintBnB.WebApi.Helpers.CurrentUserDetails;
+using static AintBnB.BusinessLogic.Helpers.Authentication;
 
 namespace AintBnB.WebApi.Controllers
 {
     [ApiController]
+    [Authorize]
     public class AccommodationController : ControllerBase
     {
         private IAccommodationService _accommodationService;
@@ -34,8 +39,14 @@ namespace AintBnB.WebApi.Controllers
             try
             {
                 var owner = await _userService.GetUserAsync(userId);
-                var newAccommodation = await _accommodationService.CreateAccommodationAsync(owner, accommodation.Address, accommodation.SquareMeters, accommodation.AmountOfBedrooms, accommodation.KilometersFromCenter, accommodation.Description, accommodation.PricePerNight, accommodation.CancellationDeadlineInDays, days);
-                return Created(HttpContext.Request.Scheme + "://" + HttpContext.Request.Host + HttpContext.Request.Path + "/" + accommodation.Id, accommodation);
+
+                if (CheckIfUserIsAllowedToPerformAction(owner, GetIdOfLoggedInUser(HttpContext)))
+                {
+                    var newAccommodation = await _accommodationService.CreateAccommodationAsync(owner, accommodation.Address, accommodation.SquareMeters, accommodation.AmountOfBedrooms, accommodation.KilometersFromCenter, accommodation.Description, accommodation.PricePerNight, accommodation.CancellationDeadlineInDays, days);
+                    return Created(HttpContext.Request.Scheme + "://" + HttpContext.Request.Host + HttpContext.Request.Path + "/" + accommodation.Id, accommodation);
+                }
+                return BadRequest($"Must be performed by a customer with ID {owner.Id}, or by admin on behalf of a customer with ID {owner.Id}!");
+
             }
             catch (Exception ex)
             {
@@ -70,6 +81,7 @@ namespace AintBnB.WebApi.Controllers
         /// <returns>Status 200 and the correct sorting order if successful, otherwise status code 404</returns>
         [HttpPost]
         [Route("api/[controller]/sort/{sortBy}/{ascOrDesc}")]
+        [IgnoreAntiforgeryToken]
         public IActionResult SortAvailableList([FromBody] List<Accommodation> available, [FromRoute] string sortBy, [FromRoute] string ascOrDesc)
         {
             try
@@ -90,15 +102,20 @@ namespace AintBnB.WebApi.Controllers
         [Route("api/[controller]/{id}/{days}")]
         public async Task<IActionResult> ExpandScheduleAsync([FromRoute] int id, [FromRoute] int days)
         {
-            try
+            var acc = await _accommodationService.GetAccommodationAsync(id);
+            if (CorrectUserOrAdmin(acc.Owner.Id, GetIdOfLoggedInUser(HttpContext), GetUsertypeOfLoggedInUser(HttpContext)))
             {
-                await _accommodationService.ExpandScheduleOfAccommodationWithXAmountOfDaysAsync(id, days);
-                return Ok(await _accommodationService.GetAccommodationAsync(id));
+                try
+                {
+                    await _accommodationService.ExpandScheduleOfAccommodationWithXAmountOfDaysAsync(id, days);
+                    return Ok(await _accommodationService.GetAccommodationAsync(id));
+                }
+                catch (Exception ex)
+                {
+                    return NotFound($"Couldn't expand the schedule for the accommodation with id {id}. {ex.Message}");
+                }
             }
-            catch (Exception ex)
-            {
-                return NotFound($"Couldn't expand the schedule for the accommodation with id {id}. {ex.Message}");
-            }
+            return NotFound($"Must be performed by a customer with ID {acc.Owner.Id}, or by admin on behalf of a customer with ID {acc.Owner.Id}!");
         }
 
         /// <summary>API PUT request to update an existing accommodation</summary>
@@ -111,8 +128,15 @@ namespace AintBnB.WebApi.Controllers
         {
             try
             {
-                await _accommodationService.UpdateAccommodationAsync(id, accommodation);
-                return Ok(accommodation);
+                var acc = await _accommodationService.GetAccommodationAsync(id);
+                var userType = GetUsertypeOfLoggedInUser(HttpContext);
+
+                if (CorrectUserOrAdmin(acc.Owner.Id, GetIdOfLoggedInUser(HttpContext), userType))
+                {
+                    await _accommodationService.UpdateAccommodationAsync(id, accommodation);
+                    return Ok(accommodation);
+                }
+                return BadRequest(new AccessException().Message);
             }
             catch (Exception ex)
             {
@@ -179,6 +203,11 @@ namespace AintBnB.WebApi.Controllers
         {
             try
             {
+                var accommodation = await _accommodationService.GetAccommodationAsync(id);
+
+                if (!CorrectUserOrAdmin(accommodation.Owner.Id, GetIdOfLoggedInUser(HttpContext), GetUsertypeOfLoggedInUser(HttpContext)))
+                    return BadRequest(new AccessException($"Administrator or user with ID {accommodation.Owner.Id} only!").Message);
+
                 await _deletionService.DeleteAccommodationAsync(id);
                 return Ok("Deletion ok");
             }
